livestock-tracker-api-gateway/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
backports.zoneinfo==0.2.1;python_version<"3.9"
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
grpcio==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
httptools==0.6.1
idna==3.6
pika==1.3.2
protobuf==4.25.3
psycopg==3.1.18
pydantic==2.6.2
pydantic-core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
